add cost_graph

+++ AMIGOS/website/our_functions/calculate_monthly_cost_without_system.py
@@ -4,5 +4,5 @@
 def calculate_monthly_cost_without_system(price_per_kW: float, monthly_user_consumption: list):
     monthly_price = []
     for each_month in monthly_user_consumption:
-        monthly_price.append(each_month * price_per_kW)
+        monthly_price.append(int(each_month * price_per_kW))
     return monthly_price
